Wire node-runtime comm state; extract SetupParts; roll out BlockHeight/SeqNum
Add create_comm_state (transport + ConnectionsCell + RPConf + CommUtil +
BlockRetriever) and extract SetupParts from setup() so the block-processing
streams can be wired next. Carry the BlockHeight/SeqNum refinements through the
block/DAG/validate/genesis layers.

+++ casper/src/block_random_seed.rs
@@ -62,7 +62,7 @@ impl BlockRandomSeed {
         } else {
             BlockRandomSeed::new(
                 block.shard_id.clone(),
-                block.block_number,
+                i64::from(block.block_number),
                 PublicKey::new(block.sender.as_bytes().to_vec()),
                 Blake2b256Hash::from_byte_array(&block.pre_state_hash),
             )
